crypto_ai_discord_bot/requirements.txt
ccxt
pandas
discord.py
python-dotenv
openai
pandas-ta
